+++ spring-jms/src/test/java/org/springframework/jms/support/JmsMessageHeaderAccessorTests.java
@@ -62,14 +62,14 @@ public class JmsMessageHeaderAccessorTests {
 		assertThat(headerAccessor.getCorrelationId()).isEqualTo("correlation-1234");
 		assertThat(headerAccessor.getDestination()).isEqualTo(destination);
 		assertThat(headerAccessor.getDeliveryMode()).isEqualTo(Integer.valueOf(1));
-		assertThat((double) headerAccessor.getExpiration()).isCloseTo((double) 1234L, within(0.0));
+		assertThat(headerAccessor.getExpiration()).isCloseTo((double) 1234L, within(0.0));
 
 		assertThat(headerAccessor.getMessageId()).isEqualTo("abcd-1234");
 		assertThat(headerAccessor.getPriority()).isEqualTo(Integer.valueOf(9));
 		assertThat(headerAccessor.getReplyTo()).isEqualTo(replyTo);
 		assertThat(headerAccessor.getRedelivered()).isEqualTo(true);
 		assertThat(headerAccessor.getType()).isEqualTo("type");
-		assertThat((double) headerAccessor.getTimestamp()).isCloseTo((double) 4567L, within(0.0));
+		assertThat(headerAccessor.getTimestamp()).isCloseTo((double) 4567L, within(0.0));
 
 		// Making sure replyChannel is not mixed with replyTo
 		assertNull(headerAccessor.getReplyChannel());
